enroll_app/api/views.py
from enroll_app.models import User_model
from enroll_app.api.serializers import UserSerializer
from rest_framework import viewsets
from rest_framework.authentication import SessionAuthentication
from rest_framework.permissions import IsAuthenticatedOrReadOnly
class UserViewSet(viewsets.ModelViewSet):
    queryset = User_model.objects.all()
    serializer_class = UserSerializer
    authentication_classes = [SessionAuthentication]
    permission_classes = [IsAuthenticatedOrReadOnly]
    

