requirements.txt
Django==4.2.13
mysqlclient==2.2.4
djangorestframework==3.15.1
gunicorn
